Plugin_Query: widen sel uint8_t->uint16_t (carry the type byte)
The plugin query selector packs (type<<8)|index (e.g. MORITZ mo_getvars). The uint8_t sel param truncated the high byte, so the type was lost and MORITZ window-contact (type 1) queries collapsed to type 0 ('Button'). Fixes both TinyC pluginQuery and Scripter mo(). Verified on the RA8876 (.102): reads all 9 window contacts.

+++ tasmota/tasmota_xdrv_driver/xdrv_10_scripter.ino
@@ -837,7 +837,7 @@ typedef struct {
 
 SCRIPT_MEM glob_script_mem;
 
-uint32_t Plugin_Query(uint16_t, uint8_t, char *);
+uint32_t Plugin_Query(uint16_t, uint16_t, char *);
 
 void script_setaflg(uint8_t flg) {
   glob_script_mem.tasm_cmd_activ = flg;
